seller controller product quantity mmangement

+++ RaghavEcomm/src/main/java/com/raghavEcomm/service/ProductServiceImpl.java
@@ -1,24 +1,44 @@
 package com.raghavEcomm.service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.raghavEcomm.exceptions.AdminException;
 import com.raghavEcomm.exceptions.LoginException;
+import com.raghavEcomm.exceptions.OrderException;
 import com.raghavEcomm.exceptions.ProductCategoryException;
 import com.raghavEcomm.exceptions.ProductException;
+import com.raghavEcomm.exceptions.UserException;
+import com.raghavEcomm.model.Cart;
 import com.raghavEcomm.model.CurrentUserSession;
+import com.raghavEcomm.model.Order;
 import com.raghavEcomm.model.Product;
 import com.raghavEcomm.model.ProductCategory;
+import com.raghavEcomm.model.Seller;
+import com.raghavEcomm.repository.CartRepo;
 import com.raghavEcomm.repository.CurrentUserSessionRepo;
+import com.raghavEcomm.repository.OrderRepo;
 import com.raghavEcomm.repository.ProductCategoryRepo;
 import com.raghavEcomm.repository.ProductRepo;
+import com.raghavEcomm.repository.SellerRepo;
 
 @Service
 public class ProductServiceImpl implements ProductService {
+
+	@Autowired
+	private SellerRepo uRepo;
+
+	@Autowired
+	private CartRepo cartRepo;
+
+	@Autowired
+	private OrderRepo orderRepo;
 
 	@Autowired
 	private ProductRepo productRepo;
@@ -31,7 +51,7 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public Product addProduct(Product product, String sellerKey, Integer categoryId)
-			throws LoginException, AdminException, ProductCategoryException {
+			throws LoginException, AdminException, ProductCategoryException, UserException {
 
 		CurrentUserSession loggedInUser = csdao.findByUuid(sellerKey);
 
@@ -47,10 +67,29 @@ public class ProductServiceImpl implements ProductService {
 
 		if (existedCategory.isPresent()) {
 
-			ProductCategory presentCategory = existedCategory.get();
-			product.setProductCategory(presentCategory);
+			Optional<Seller> existingSeller = uRepo.findById(loggedInUser.getUserId());
 
-			return productRepo.save(product);
+			if (existingSeller.isPresent()) {
+
+				Seller seller = existingSeller.get();
+
+				ProductCategory presentCategory = existedCategory.get();
+				product.setProductCategory(presentCategory);
+
+				Map<Product, Integer> map = seller.getProducts();
+
+				if (map.containsKey(product)) {
+					map.put(product, map.get(product) + product.getProductQuantity());
+					product.setProductQuantity(product.getProductQuantity() + map.get(product));
+				} else {
+					map.put(product, product.getProductQuantity());
+				}
+
+				return productRepo.save(product);
+
+			} else {
+				throw new UserException("Seller Not Found");
+			}
 
 		} else {
 			throw new ProductCategoryException("Category does not exists with this category Id" + categoryId);
@@ -60,7 +99,7 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public Product updateProduct(Product product, String sellerKey, Integer productId)
-			throws LoginException, AdminException, ProductException {
+			throws LoginException, AdminException, ProductException, UserException {
 
 		CurrentUserSession loggedInUser = csdao.findByUuid(sellerKey);
 
@@ -80,7 +119,26 @@ public class ProductServiceImpl implements ProductService {
 
 			product.setProductCategory(present.getProductCategory());
 
-			return productRepo.save(product);
+			Optional<Seller> existingSeller = uRepo.findById(loggedInUser.getUserId());
+
+			if (existingSeller.isPresent()) {
+
+				Seller seller = existingSeller.get();
+
+				Map<Product, Integer> map = seller.getProducts();
+
+				if (map.containsKey(product)) {
+					map.put(product, product.getProductQuantity());
+				} else {
+
+					throw new ProductException("Product not associated with seller");
+				}
+
+				return productRepo.save(product);
+
+			} else {
+				throw new UserException("Seller Not Found");
+			}
 
 		} else {
 			throw new ProductException("Product does not exists with this product Id" + productId);
@@ -89,7 +147,7 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public Product deleteProduct(String sellerKey, Integer productId)
-			throws ProductException, LoginException, AdminException {
+			throws ProductException, LoginException, AdminException, OrderException {
 
 		CurrentUserSession loggedInUser = csdao.findByUuid(sellerKey);
 
@@ -109,9 +167,36 @@ public class ProductServiceImpl implements ProductService {
 
 			present.setProductCategory(null);
 
-			productRepo.delete(present);
+			if (!getProductInOrders(present)) {
 
-			return present;
+				List<Cart> carts = getAllCartWithProduct(present);
+
+				if (!carts.isEmpty()) {
+
+					for (Cart c : carts) {
+						c.getProducts().remove(present);
+					}
+
+				}
+
+				Optional<Seller> currentSeller = uRepo.findById(loggedInUser.getUserId());
+
+				if (currentSeller.isPresent()) {
+					Seller seller = currentSeller.get();
+
+					seller.getProducts().remove(present);
+
+					productRepo.delete(present);
+
+					return present;
+
+				} else {
+					throw new AdminException("Seller not Found");
+				}
+
+			} else {
+				throw new OrderException("Product can't deleted once order is placed");
+			}
 
 		} else {
 			throw new ProductException("Product does not exists with this product Id" + productId);
@@ -165,6 +250,40 @@ public class ProductServiceImpl implements ProductService {
 		}
 
 		return list;
+	}
+
+	@Override
+	public List<Cart> getAllCartWithProduct(Product product) {
+
+		List<Cart> carts = new ArrayList<>();
+
+		carts = cartRepo.findAll();
+
+		if (carts.isEmpty())
+			return carts;
+
+		carts = carts.stream().filter(c -> c.getProducts().containsKey(product)).collect(Collectors.toList());
+
+		return carts;
+
+	}
+
+	@Override
+	public boolean getProductInOrders(Product product) {
+
+		List<Order> ordersList = orderRepo.findAll();
+
+		if (ordersList == null)
+			return false;
+
+		List<Order> newList = ordersList.stream().filter(o -> o.getProducts().containsKey(product))
+				.collect(Collectors.toList());
+
+		if (newList.size() > 0) {
+			return true;
+		} else {
+			return false;
+		}
 	}
 
 }
